subtree : fix split output containing unrelated history
When a subtree is deleted and re-added at the same location with the
--squash arguement the history created by "git subtree split" will
contain commits that do not affect the files in the given prefix
resulting in an incorrect history for the subtree.

This commit fixes the issue by removing lines that create cache files for
commits that did not contain the given prefix after new parent commits have
been created. The history created by "git subtree split" with this change
no longer contains unrelated commits or commits that occured before the
subtree was removed and re-added.

Signed-off-by: Taylor Jones <tjones629@gmail.com>