gitk: tag add right click options
In gitk, we can right-click on the icon of the branch,
and a directory will pop up to provide us with functions
such as "Checkout this branch", "Rename this branch"...,
but we found that the right-click tag icon does not have
such a function , So I learned how to write the branch icon,
and added the following functions "Rename this tag",
"Remove this tag", "Copy tag name" to right-click the tag icon.

This function temporarily only supports those labels independently
displayed tag(s), and cannot work on label aggregated display
"tag..." or "n tags...".

Signed-off-by: ZheNing Hu <adlternative@gmail.com>